backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
google-cloud-storage
python-pptx
pytest
